+++ packages/core/src/components/index.ts
@@ -1,3 +1,4 @@
 export * from './icons';
 export * from './bento-button';
 export * from './bento-switch';
+export * from './textbox';
